fix: Correct bottom space for "Book Bar Only"

+++ src/activities/reader/TxtReaderActivity.h
@@ -26,7 +26,7 @@ class TxtReaderActivity final : public ActivityWithSubactivity {
 
   // Cached settings for cache validation (different fonts/margins require re-indexing)
   int cachedFontId = 0;
-  int cachedScreenMargin = 0;
+  uint8_t cachedScreenMargin = 0;
   uint8_t cachedParagraphAlignment = CrossPointSettings::LEFT_ALIGN;
 
   void renderPage();
